+++ src/dataops/plugin/plugin_manager.py
@@ -16,7 +16,7 @@ from django.utils.dateparse import parse_datetime
 from django.utils.translation import ugettext, ugettext_lazy as _
 
 from dataops import settings
-from dataops.models import PluginRegistry
+from dataops.models import Plugin
 from dataops.pandas import load_table, perform_dataframe_upload_merge
 from dataops.plugin import ontask_plugin
 from dataops.plugin.ontask_plugin import OnTaskPluginAbstract
@@ -129,7 +129,7 @@ def _verify_plugin(pinobj):
 
         # Verify that all the fields and methods are present in the instance
         diag[check_idx] = _('Not found')
-        if pinobj.name is not None and isinstance(pinobj.name, str):
+        if pinobj.name and isinstance(pinobj.name, str):
             diag[check_idx] = _('Ok')
         else:
             diag[check_idx] = _('Incorrect type')
@@ -235,7 +235,7 @@ def _verify_plugin(pinobj):
 
 
 def _load_plugin_info(plugin_folder, plugin_rego=None):
-    """Load the plugin and populate the PluginRegistry table.
+    """Load the plugin and populate the Plugin table.
 
     :param plugin_folder: Folder to load the information from.
 
@@ -249,7 +249,7 @@ def _load_plugin_info(plugin_folder, plugin_rego=None):
 
     # If there is no instance given of the registry, create a new one
     if not plugin_rego:
-        plugin_rego = PluginRegistry()
+        plugin_rego = Plugin()
         plugin_rego.filename = plugin_folder
 
     if plugin_instance:
@@ -261,6 +261,8 @@ def _load_plugin_info(plugin_folder, plugin_rego=None):
             plugin_rego.is_model = False
 
     plugin_rego.is_verified = plugin_instance is not None
+    if not plugin_rego.is_verified:
+        plugin_rego.is_enabled = False
 
     # All went good
     plugin_rego.save()
@@ -300,11 +302,11 @@ def load_plugin(foldername):
     return plugin_instance, tests
 
 
-def refresh_plugin_data(request, workflow):
+def refresh_plugin_data(request, workflow = None):
     """Refresh the plugin data in the system.
 
     Function to traverse the directory where the plugins live and check if
-    the folders in there are reflected in the PluginRegistry model.
+    the folders in there are reflected in the Plugin model.
 
     :return: Reflect the changes in the database
     """
@@ -316,9 +318,9 @@ def refresh_plugin_data(request, workflow):
     ]
 
     # Get the objects from the DB
-    reg_plugins = PluginRegistry.objects.all()
+    reg_plugins = Plugin.objects.all()
 
-    # Travers the list of registered plugins and detect changes
+    # Traverse the list of registered plugins and detect changes
     for rpin in reg_plugins:
         i_file = os.path.join(plugin_folder, rpin.filename, '__init__.py')
         if rpin.filename not in pfolders or not os.path.exists(i_file):