+++ src/dataops/views/transform.py
@@ -1,41 +1,38 @@
 # -*- coding: utf-8 -*-
 
-"""Views to manipulate the transformation."""
+"""Views to manipulate the transformations and models."""
 
 import json
 from builtins import object, str
-from typing import Optional
+from typing import Dict, List, Optional, Tuple
 
 import django_tables2 as tables
-from celery.task.control import inspect
 from django.contrib import messages
 from django.contrib.auth.decorators import user_passes_test
 from django.db.models import F
-from django.http import HttpRequest, HttpResponse, JsonResponse
+from django.http import HttpRequest, HttpResponse
 from django.shortcuts import redirect, render, reverse
-from django.template.loader import render_to_string
 from django.urls import resolve
 from django.utils.html import format_html
 from django.utils.translation import ugettext_lazy as _
 
 from dataops.forms import FIELD_PREFIX, PluginInfoForm
-from dataops.models import PluginRegistry
+from dataops.models import Plugin
 from dataops.plugin.plugin_manager import load_plugin, refresh_plugin_data
 from logs.models import Log
-from ontask.decorators import ajax_required, get_workflow
+from ontask.celery import celery_is_up
+from ontask.decorators import get_workflow
 from ontask.permissions import is_instructor
 from ontask.tasks import run_plugin_task
 from workflow.models import Workflow
 
 
-class PluginRegistryTable(tables.Table):
+class PluginAvailableTable(tables.Table):
     """Class to render the table with plugins available for execution.
 
     The Operations column is inheriting from another class to centralise the
     customisation.
     """
-
-    name = tables.Column(verbose_name=_('Name'), empty_values=None)
 
     description_txt = tables.TemplateColumn(
         verbose_name=_('Description'),
@@ -45,7 +42,7 @@ class PluginRegistryTable(tables.Table):
     last_exec = tables.DateTimeColumn(verbose_name=_('Last executed'))
 
     def __init__(self, *args, **kwargs):
-        """Set workflow and user to get latest execution time"""
+        """Set workflow and user to get latest execution time."""
         self.workflow = kwargs.pop('workflow')
         self.user = kwargs.pop('user')
 
@@ -81,9 +78,9 @@ class PluginRegistryTable(tables.Table):
         return log_item.created
 
     class Meta(object):
-        """Choose model, fields, sequence and attributes."""
+        """Choose fields, sequence and attributes."""
 
-        model = PluginRegistry
+        model = Plugin
 
         fields = ('name', 'description_txt')
 
@@ -94,6 +91,67 @@ class PluginRegistryTable(tables.Table):
             'style': 'width: 100%;',
             'id': 'transform-table',
         }
+
+
+def _prepare_plugin_input_output(
+    plugin_instance: Plugin,
+    workflow: Workflow,
+    form: PluginInfoForm,
+) -> Tuple[List, List]:
+    """Prepare input, output and parameters for plugin execution.
+
+    :param plugin_instance: Plugin object allowed to execute
+
+    :param workflow: Workflow object (to access columns and similar)
+
+    :param form: Form just populated with a POST
+
+    :return: [list of inputs, list of outputs]
+    """
+    # Take the list of inputs from the form if empty list is given.
+    input_column_names = []
+    if plugin_instance.input_column_names:
+        # Traverse the fields and get the names of the columns
+        for idx, __ in enumerate(plugin_instance.input_column_names):
+            input_column_names.append(
+                workflow.columns.get(id=int(
+                    form.cleaned_data[FIELD_PREFIX + 'input_%s' % idx],
+                )).name,
+            )
+    else:
+        input_column_names = [col.name for col in form.cleaned_data['columns']]
+
+    output_column_names = []
+    if plugin_instance.output_column_names:
+        # Process the output columns
+        for idx, __ in enumerate(plugin_instance.output_column_names):
+            output_column_names.append(
+                form.cleaned_data[FIELD_PREFIX + 'output_%s' % idx])
+    else:
+        # Plugin instance has an empty set of output files, clone the input
+        output_column_names = input_column_names[:]
+
+    suffix = form.cleaned_data['out_column_suffix']
+    if suffix:
+        # A suffix has been provided, add it to the list of outputs
+        output_column_names = [
+            cname + suffix for cname in output_column_names
+        ]
+
+    return input_column_names, output_column_names
+
+
+def _prepare_plugin_parameters(
+    plugin_instance: Plugin,
+    form: PluginInfoForm,
+) -> Dict:
+    # Pack the parameters
+    exec_params = {}
+    for idx, tpl in enumerate(plugin_instance.parameters):
+        exec_params[tpl[0]] = form.cleaned_data[
+            FIELD_PREFIX + 'parameter_%s' % idx]
+
+    return exec_params
 
 
 @user_passes_test(is_instructor)
@@ -117,62 +175,23 @@ def transform_model(
     # Traverse the plugin folder and refresh the db content.
     refresh_plugin_data(request, workflow)
 
-    table_ok = PluginRegistryTable(
-        PluginRegistry.objects.filter(
+    table_ok = PluginAvailableTable(
+        Plugin.objects.filter(
             is_model=is_model,
             is_verified=True,
-        ),
+            is_enabled=True),
         orderable=False,
         user=request.user,
         workflow=workflow)
 
     table_err = None
     if request.user.is_superuser:
-        table_err = PluginRegistry.objects.filter(is_model=None)
+        table_err = Plugin.objects.filter(is_model=None)
 
     return render(
         request,
         'dataops/transform_model.html',
         {'table': table_ok, 'is_model': is_model, 'table_err': table_err})
-
-
-@user_passes_test(is_instructor)
-@ajax_required
-def diagnose(
-    request: HttpRequest,
-    pk: int,
-    workflow: Optional[Workflow] = None,
-) -> JsonResponse:
-    """Show the diagnostics of a plugin that failed the verification tests.
-
-    :param request: HTML request object
-
-    :param pk: Primary key of the transform element
-
-    :return:
-    """
-    # Action being used
-    plugin = PluginRegistry.objects.filter(id=pk).first()
-    if not plugin:
-        return JsonResponse({'html_redirect': reverse('home')})
-
-    # Reload the plugin to get the messages stored in the right place.
-    pinstance, msgs = load_plugin(plugin.filename)
-
-    # If the new instance is now properly verified, simply redirect to the
-    # transform page
-    if pinstance:
-        plugin.is_verified = True
-        plugin.save()
-        return JsonResponse({'html_redirect': reverse('dataops:transform')})
-
-    # Get the diagnostics from the plugin and use it for rendering.
-    return JsonResponse({
-        'html_form': render_to_string(
-            'dataops/includes/partial_diagnostics.html',
-            {'diagnostic_table': msgs, 'folder': plugin.filename},
-            request=request),
-    })
 
 
 @user_passes_test(is_instructor)
@@ -188,22 +207,16 @@ def plugin_invoke(
     :param pk: primary key of the plugin
     :return: Page offering to select the columns to invoke
     """
-
     # Verify that celery is running!
-    celery_stats = None
-    try:
-        celery_stats = inspect().stats()
-    except Exception:
-        # If the stats are empty, celery is not running.
-        if not celery_stats:
-            messages.error(
-                request,
-                _(
-                    'Unable to run plugins due to a misconfiguration. '
-                    + 'Ask your system administrator to enable queueing.'))
-            return redirect(reverse('table:display'))
+    if not celery_is_up():
+        messages.error(
+            request,
+            _(
+                'Unable to run plugins due to a misconfiguration. '
+                + 'Ask your system administrator to enable queueing.'))
+        return redirect(reverse('table:display'))
 
-    plugin_info = PluginRegistry.objects.filter(pk=pk).first()
+    plugin_info = Plugin.objects.filter(pk=pk).first()
     if not plugin_info:
         return redirect('home')
 
@@ -214,7 +227,7 @@ def plugin_invoke(
             {'empty_wflow': True,
              'is_model': plugin_info.get_is_model()})
 
-    plugin_instance, msgs = load_plugin(plugin_info.filename)
+    plugin_instance, __ = load_plugin(plugin_info.filename)
     if plugin_instance is None:
         messages.error(
             request,
@@ -228,58 +241,15 @@ def plugin_invoke(
         workflow=workflow,
         plugin_instance=plugin_instance)
 
-    # Set the basic elements in the context
-    context = {
-        'form': form,
-        'input_column_fields': [
-            fld for fld in list(form)
-            if fld.name.startswith(FIELD_PREFIX + 'input')],
-        'output_column_fields': [
-            fld for fld in list(form)
-            if fld.name.startswith(FIELD_PREFIX + 'output')],
-        'parameters': [
-            fld for fld in list(form)
-            if fld.name.startswith(FIELD_PREFIX + 'parameter')],
-        'pinstance': plugin_instance,
-        'id': workflow.id,
-    }
-
     if request.method == 'POST' and form.is_valid():
-        # Take the list of inputs from the form if empty list is given.
-        input_column_names = []
-        if plugin_instance.input_column_names:
-            # Traverse the fields and get the names of the columns
-            for idx, __ in enumerate(plugin_instance.input_column_names):
-                cid = form.cleaned_data[FIELD_PREFIX + 'input_%s' % idx]
-                input_column_names.append(
-                    workflow.columns.get(id=int(cid)).name,
-                )
-        else:
-            input_column_names = [
-                col.name for col in form.cleaned_data['columns']]
+        in_cols, out_cols = _prepare_plugin_input_output(
+            plugin_instance,
+            workflow,
+            form)
 
-        output_column_names = []
-        if plugin_instance.output_column_names:
-            # Process the output columns
-            for idx, __ in enumerate(plugin_instance.output_column_names):
-                new_cname = form.cleaned_data[FIELD_PREFIX + 'output_%s' % idx]
-                output_column_names.append(new_cname)
-        else:
-            # Plugin instance has an empty set of output files, clone the input
-            output_column_names = input_column_names[:]
-
-        suffix = form.cleaned_data['out_column_suffix']
-        if suffix:
-            # A suffix has been provided, add it to the list of outputs
-            output_column_names = [
-                cname + suffix for cname in output_column_names
-            ]
-
-        # Pack the parameters
-        parameters = {}
-        for idx, tpl in enumerate(plugin_instance.parameters):
-            parameters[tpl[0]] = form.cleaned_data[
-                FIELD_PREFIX + 'parameter_%s' % idx]
+        exec_params = _prepare_plugin_parameters(
+            plugin_instance,
+            form)
 
         # Log the event with the status "preparing invocation"
         log_item = Log.objects.register(
@@ -289,9 +259,9 @@ def plugin_invoke(
             {
                 'id': plugin_info.id,
                 'name': plugin_info.name,
-                'input_column_names': input_column_names,
-                'output_column_names': output_column_names,
-                'parameters': json.dumps(parameters, default=str),
+                'input_column_names': in_cols,
+                'output_column_names': out_cols,
+                'parameters': json.dumps(exec_params, default=str),
                 'status': 'preparing execution'})
 
         run_plugin_task.apply_async(
@@ -299,11 +269,11 @@ def plugin_invoke(
                 request.user.id,
                 workflow.id,
                 pk,
-                input_column_names,
-                output_column_names,
-                suffix,
+                in_cols,
+                out_cols,
+                form.cleaned_data['out_column_suffix'],
                 form.cleaned_data['merge_key'],
-                parameters,
+                exec_params,
                 log_item.id),
             serializer='pickle')
 
@@ -316,37 +286,17 @@ def plugin_invoke(
     return render(
         request,
         'dataops/plugin_info_for_run.html',
-        context)
-
-
-@user_passes_test(is_instructor)
-@ajax_required
-@get_workflow()
-def moreinfo(
-    request: HttpRequest,
-    pk: int,
-    workflow: Optional[Workflow] = None,
-) -> JsonResponse:
-    """Show the detailed information about a plugin.
-
-    :param request: HTML request object
-
-    :param pk: Primary key of the PluginRegistry element
-
-    :return:
-    """
-    # Action being used
-    plugin = PluginRegistry.objects.filter(id=pk).first()
-    if not plugin:
-        return JsonResponse({'html_redirect': reverse('home')})
-
-    # Reload the plugin to get the messages stored in the right place.
-    pinstance, msgs = load_plugin(plugin.filename)
-
-    # Get the descriptions and show them in the modal
-    return JsonResponse({
-        'html_form': render_to_string(
-            'dataops/includes/partial_plugin_long_description.html',
-            {'pinstance': pinstance},
-            request=request),
-    })
+        context={
+            'form': form,
+            'input_column_fields': [
+                fld for fld in list(form)
+                if fld.name.startswith(FIELD_PREFIX + 'input')],
+            'output_column_fields': [
+                fld for fld in list(form)
+                if fld.name.startswith(FIELD_PREFIX + 'output')],
+            'parameters': [
+                fld for fld in list(form)
+                if fld.name.startswith(FIELD_PREFIX + 'parameter')],
+            'pinstance': plugin_instance,
+            'id': workflow.id,
+        })