+++ src/logs/models.py
@@ -49,6 +49,7 @@ class Log(models.Model):
     WORKFLOW_IMPORT = 'workflow_import'
     WORKFLOW_CLONE = 'workflow_clone'
     WORKFLOW_UPDATE_LUSERS = 'workflow_update_lusers'
+    WORKFLOW_STAR = 'workflow_star'
     COLUMN_ADD = 'column_add'
     QUESTION_ADD = 'question_add'
     COLUMN_ADD_FORMULA = 'column_add_formula'
@@ -118,6 +119,7 @@ class Log(models.Model):
         (WORKFLOW_IMPORT, _('Import workflow')),
         (WORKFLOW_CLONE, _('Workflow cloned')),
         (WORKFLOW_UPDATE_LUSERS, _('Update list of workflow users')),
+        (WORKFLOW_STAR, _('Toggle workflow star')),
         (COLUMN_ADD, _('Column added')),
         (COLUMN_ADD_FORMULA, _('Column with formula created')),
         (COLUMN_ADD_RANDOM, _('Column with random values created')),