+++ src/workflow/views/workflow_ops.py
@@ -22,8 +22,9 @@ from dataops.sql import get_rows, get_text_column_hash
 from logs.models import Log
 from ontask import is_correct_email
 from ontask.decorators import (
-    ajax_required, get_column, get_workflow, store_workflow_in_session,
+    ajax_required, get_column, get_workflow,
 )
+from ontask.workflow_access import store_workflow_in_session
 from ontask.permissions import is_instructor
 from ontask.tables import OperationsColumn
 from ontask.tasks import workflow_update_lusers
@@ -171,6 +172,35 @@ def flush(
             {'workflow': workflow},
             request=request),
     })
+
+
+@user_passes_test(is_instructor)
+@ajax_required
+@get_workflow()
+def star(
+    request: HttpRequest,
+    wid: Optional[int] = None,
+    workflow: Optional[Workflow] = None,
+) -> JsonResponse:
+    """Toggle the star mark in the workflow."""
+    # Get the workflows with stars
+    stars = request.user.workflows_star.all()
+    if workflow in stars:
+        workflow.star.remove(request.user)
+        has_star = False
+    else:
+        workflow.star.add(request.user)
+        has_star = True
+
+    # Log the event
+    Log.objects.register(
+        request.user,
+        Log.WORKFLOW_STAR,
+        workflow,
+        {'id': workflow.id, 'name': workflow.name, 'star': has_star})
+
+    # In this case, the form is valid
+    return JsonResponse({})
 
 
 @user_passes_test(is_instructor)